voice_bot/requirements.txt
fastapi
uvicorn
jinja2
speechrecognition
pyttsx3
python-dotenv
groq
ffmpeg-python
pydantic
